requirements.txt
fastapi>=0.110
uvicorn>=0.27
google-genai>=0.3.0
